# swagger.json
{"openapi":"3.0.0","paths":{"/api/v1":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User registered successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}},"400":{"description":"Email already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Register a new user","tags":["Authentication"]}},"/api/v1/auth/verify-otp":{"post":{"operationId":"AuthController_verifyOTP","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOtpDto"}}}},"responses":{"200":{"description":"Email verified successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}},"400":{"description":"Invalid or expired OTP","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Verify email with OTP","tags":["Authentication"]}},"/api/v1/auth/resend-otp":{"post":{"operationId":"AuthController_resendOTP","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendOtpDto"}}}},"responses":{"200":{"description":"OTP resent successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Resend verification OTP","tags":["Authentication"]}},"/api/v1/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"200":{"description":"Password reset OTP sent successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Request password reset OTP","tags":["Authentication"]}},"/api/v1/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":"Password reset successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}},"404":{"description":"Invalid or expired OTP","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Reset password using OTP","tags":["Authentication"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"User logged in successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"User login","tags":["Authentication"]}},"/api/v1/auth/google":{"post":{"operationId":"AuthController_googleLogin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoogleLoginDto"}}}},"responses":{"200":{"description":"Google login successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}},"401":{"description":"Google authentication failed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Google OAuth login/register","tags":["Authentication"]}},"/api/v1/auth/profile":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":"Returns access confirmation","content":{"application/json":{"schema":{"example":{"message":"AuthGuard and RolesGuard are working!"}}}}},"401":{"description":"Missing or invalid token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"403":{"description":"Role not authorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get profile (Protected Route Test)","tags":["Authentication"]}},"/api/v1/cloudinary/signature":{"get":{"operationId":"CloudinaryController_getSignature","parameters":[{"name":"folder","required":false,"in":"query","description":"Cloudinary folder","schema":{"example":"avatars","type":"string"}}],"responses":{"200":{"description":"Returns upload signature and configuration","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloudinarySignatureResponseDto"}}}},"401":{"description":"Unauthorized access","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Generate Cloudinary upload signature","tags":["Cloudinary"]}},"/api/v1/admin/login":{"post":{"operationId":"AdminController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminLoginDto"}}}},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminResponseDto"}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Admin login","tags":["Admin Dashboard"]}},"/api/v1/profile/me":{"get":{"operationId":"ProfileController_getProfile","parameters":[],"responses":{"200":{"description":"Profile retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileResponseDto"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get current user profile","tags":["Profile"]}},"/api/v1/profile/update":{"patch":{"operationId":"ProfileController_updateProfile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileDto"}}}},"responses":{"200":{"description":"Profile updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileResponseDto"}}}},"400":{"description":"Bad Request - Validation failed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"404":{"description":"User not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Update user profile","tags":["Profile"]}},"/api/v1/hostels":{"post":{"operationId":"HostelController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateHostelDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Create a new hostel","tags":["Hostels"]}},"/api/v1/hostels/{id}":{"patch":{"operationId":"HostelController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateHostelDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Update hostel details","tags":["Hostels"]},"get":{"operationId":"HostelController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get a single hostel by ID","tags":["Hostels"]},"delete":{"operationId":"HostelController_deleteHostel","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":"Hostel deleted successfully"},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Delete a hostel","tags":["Hostels"]}},"/api/v1/hostels/my-hostels":{"get":{"operationId":"HostelController_getMyHostels","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"tags":["Hostels"]}},"/api/v1/hostels/{hostelId}/rooms":{"post":{"operationId":"HostelController_createRoom","parameters":[{"name":"hostelId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoomDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Create a new room","tags":["Hostels"]},"get":{"operationId":"HostelController_getRoomsByHostel","parameters":[{"name":"hostelId","required":true,"in":"path","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get all rooms in a hostel","tags":["Hostels"]}},"/api/v1/hostels/{hostelId}/rooms/{roomId}":{"get":{"operationId":"HostelController_getSingleRoom","parameters":[{"name":"hostelId","required":true,"in":"path","schema":{"type":"number"}},{"name":"roomId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get single room details","tags":["Hostels"]},"patch":{"operationId":"HostelController_updateRoom","parameters":[{"name":"hostelId","required":true,"in":"path","schema":{"type":"number"}},{"name":"roomId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoomDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Update room details","tags":["Hostels"]},"delete":{"operationId":"HostelController_deleteRoom","parameters":[{"name":"hostelId","required":true,"in":"path","schema":{"type":"number"}},{"name":"roomId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":"Room deleted"},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Delete a room","tags":["Hostels"]}},"/api/v1/hostels/bookings":{"get":{"operationId":"HostelController_getOwnerBookings","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get all bookings for all hostels owned by the user","tags":["Hostels"]}},"/api/v1/student/bookings":{"post":{"operationId":"StudentController_bookRoom","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookingDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Book a room","tags":["Student"]},"get":{"operationId":"StudentController_getMyBookings","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Bookings retrieved successfully"},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]},{"bearer":[]}],"summary":"Get my bookings","tags":["Student"]}},"/api/v1/student/bookings/{bookingId}/cancel":{"patch":{"operationId":"StudentController_cancelBooking","parameters":[{"name":"bookingId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Booking cancelled successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]},{"bearer":[]}],"summary":"Cancel a pending booking","tags":["Student"]}},"/api/v1/student/hostels":{"get":{"operationId":"StudentController_getAvailableHostels","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Hostels retrieved successfully"},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]},{"bearer":[]}],"summary":"Get all available hostels","tags":["Student"]}},"/api/v1/student/rooms":{"get":{"operationId":"StudentController_getAvailableRooms","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"sort","required":false,"in":"query","schema":{"enum":["price_asc","price_desc"],"type":"string"}},{"name":"capacity","required":false,"in":"query","schema":{"type":"number"}},{"name":"price","required":false,"in":"query","schema":{"type":"number"}},{"name":"hostel_id","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Rooms retrieved successfully"},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"bearer":[]},{"bearer":[]}],"summary":"Get available rooms with filters","tags":["Student"]}}},"info":{"title":"Student Welfare Platform API","description":"The API documentation for the Student Welfare Platform","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"},"email":{"type":"string","example":"john.doe@example.com"},"password":{"type":"string","example":"P@ssw0rd"},"phone":{"type":"string","example":"+1234567890"},"image":{"type":"string","format":"binary"},"role":{"type":"string","example":"student","enum":["student","hostelOwner"]}},"required":["firstName","lastName","email","password","phone","image","role"]},"UserData":{"type":"object","properties":{"id":{"type":"number","example":1},"email":{"type":"string","example":"user@example.com"},"role":{"type":"string","example":"student"},"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"},"image":{"type":"string","example":"https://res.cloudinary.com/path/to/image.jpg","nullable":true}},"required":["id","email","role","firstName","lastName","image"]},"AuthResponseDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserData"},"token":{"type":"string","example":"eyJhbGciOiJIUzI1Ni..."}},"required":["user","token"]},"ErrorResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":false,"description":"Indicates if the request was successful"},"message":{"type":"string","example":"Error message description","description":"The error message returned by the server"},"statusCode":{"type":"number","example":400,"description":"The HTTP status code (e.g., 400, 401, 403, 500)"}},"required":["success","message","statusCode"]},"VerifyOtpDto":{"type":"object","properties":{"email":{"type":"string","example":"john.doe@example.com"},"otp_code":{"type":"string","example":"123456"}},"required":["email","otp_code"]},"MessageResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Operation successful"},"data":{"type":"object","example":null}},"required":["success","message","data"]},"ResendOtpDto":{"type":"object","properties":{"email":{"type":"string","example":"john.doe@example.com"}},"required":["email"]},"ForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string","example":"john.doe@example.com"}},"required":["email"]},"ResetPasswordDto":{"type":"object","properties":{"email":{"type":"string","example":"john.doe@example.com"},"otp_code":{"type":"string","example":"123456"},"newPassword":{"type":"string","example":"NewP@ssw0rd123"}},"required":["email","otp_code","newPassword"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"john.doe@example.com"},"password":{"type":"string","example":"P@ssw0rd"}},"required":["email","password"]},"GoogleLoginDto":{"type":"object","properties":{"idToken":{"type":"string","example":"eyJhbGciOiJSUzI1NiIs...","description":"Google ID Token"},"role":{"type":"string","enum":["student","hostelOwner"],"example":"student"},"phone":{"type":"string","example":"08012345678"}},"required":["idToken","role"]},"CloudinarySignatureResponseDto":{"type":"object","properties":{"signature":{"type":"string","example":"abc123def456..."},"timestamp":{"type":"number","example":1715432100},"cloudName":{"type":"string","example":"your-cloud-name"},"apiKey":{"type":"string","example":"1234567890"},"folder":{"type":"string","example":"avatars"}},"required":["signature","timestamp","cloudName","apiKey","folder"]},"AdminLoginDto":{"type":"object","properties":{"email":{"type":"string","example":"admin@hostel.com"},"password":{"type":"string","example":"Admin@123"}},"required":["email","password"]},"AdminData":{"type":"object","properties":{"id":{"type":"number","example":1},"email":{"type":"string","example":"admin@hostel.com"},"role":{"type":"string","example":"superadmin"}},"required":["id","email","role"]},"AdminResponseDto":{"type":"object","properties":{"message":{"type":"string","example":"Admin login successful"},"accessToken":{"type":"string","example":"eyJhbGciOiJIUzI1Ni..."},"admin":{"$ref":"#/components/schemas/AdminData"}},"required":["message","accessToken","admin"]},"UserProfileResponseDto":{"type":"object","properties":{"id":{"type":"number","example":1},"firstName":{"type":"string","example":"Super"},"lastName":{"type":"string","example":"Admin"},"email":{"type":"string","example":"admin@hostel.com"},"phone":{"type":"string","example":"+1234567890"},"image":{"type":"string","example":"https://res.cloudinary.com/path"},"role":{"type":"string","example":"superadmin"},"createdAt":{"format":"date-time","type":"string","example":"2022-01-01T00:00:00.000Z"},"updatedAt":{"format":"date-time","type":"string","example":"2022-01-01T00:00:00.000Z"}},"required":["id","firstName","lastName","email","phone","image","role","createdAt","updatedAt"]},"UpdateProfileDto":{"type":"object","properties":{"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"},"phone":{"type":"string","example":"+1234567890"},"image":{"type":"string","example":"https://example.com/profile.jpg"},"address":{"type":"string","example":"123 Main St"}}},"CreateHostelResourceDto":{"type":"object","properties":{"file_url":{"type":"string","example":"https://res.cloudinary.com/..."},"file_type":{"type":"string","enum":["IMAGE","VIDEO","PDF"],"example":"IMAGE"}},"required":["file_url","file_type"]},"CreateHostelDto":{"type":"object","properties":{"name":{"type":"string","example":"Indigo Hostel"},"location":{"type":"string","example":"123 Campus Road, Tokyo"},"description":{"type":"string","example":"A serene environment for students with 24/7 power."},"resources":{"type":"array","items":{"$ref":"#/components/schemas/CreateHostelResourceDto"}}},"required":["name","location","description"]},"UpdateHostelDto":{"type":"object","properties":{"name":{"type":"string","example":"Indigo Hostel"},"location":{"type":"string","example":"123 Campus Road, Tokyo"},"description":{"type":"string","example":"A serene environment for students with 24/7 power."},"resources":{"type":"array","items":{"$ref":"#/components/schemas/CreateHostelResourceDto"}}}},"CreateRoomResourceDto":{"type":"object","properties":{"file_url":{"type":"string","example":"https://res.cloudinary.com/..."},"file_type":{"type":"string","enum":["IMAGE","VIDEO","PDF"],"example":"IMAGE"}},"required":["file_url","file_type"]},"CreateRoomDto":{"type":"object","properties":{"room_number":{"type":"string","example":"Room 101"},"capacity":{"type":"number","example":2},"price":{"type":"number","example":50000},"availability":{"type":"boolean","example":true},"resources":{"type":"array","items":{"$ref":"#/components/schemas/CreateRoomResourceDto"}}},"required":["room_number","capacity","price"]},"UpdateRoomDto":{"type":"object","properties":{"room_number":{"type":"string","example":"Room 101"},"capacity":{"type":"number","example":2},"price":{"type":"number","example":50000},"availability":{"type":"boolean","example":true}}},"CreateBookingDto":{"type":"object","properties":{"room_id":{"type":"number","example":101},"reference":{"type":"string","example":"REF-12345"}},"required":["room_id","reference"]}}}}